feat(import): accept more actions for existing imports and return status
Only accepting a boolean flag to overwrite pending imports is no good
long term solution because a future flag to update already accepted
imports only makes sense if pending imports are also updated.
To account for all the possible outcomes of an import, the function
should return the import's status in addition to the ID.
Not all possible values will be supported by the initial version.