chore(eslint): Drop comma-dangle rule
The rule to disallow trailing commas often forces you to pointlessly
touch neighboring lines when you add or remove items.

If we want to enforce consistent dangling commas, we should rather force
them to always be present. But since our whole codebase follows this
rule, it would be better to have a grace period to gradually add commas.

https://github.com/metabrainz/bookbrainz-site/commit/2fc11a22d2750b320fe48dc6a15ccab393137472